0.9

+++ backend/app/controllers/portfolio_items_controller.rb
@@ -9,12 +9,19 @@ class PortfolioItemsController < ApplicationController
       end
     
     def create
+      pi = PortfolioItem.find_by(ticker: portfolio_item_params[:ticker], user_id: portfolio_item_params[:user_id])
+      if !pi.nil?
+        pi.update(quantity: (portfolio_item_params[:quantity].to_i + pi[:quantity].to_i))
+        render json: pi
+        return
+      else
         portfolio_item = PortfolioItem.create(portfolio_item_params)
         if portfolio_item.valid?
           render json: portfolio_item
         else
           render json: { errors: portfolio_item.errors.full_messages }, status: :unprocessable_entity
         end
+      end 
       end
     
       def update
@@ -25,8 +32,14 @@ class PortfolioItemsController < ApplicationController
     
       def destroy
         portfolio_item = PortfolioItem.find(params.require(:id))
-        portfolio_item.destroy
-        render json: portfolio_item
+        if portfolio_item[:quantity].to_i > portfolio_item_params[:quantity].to_i
+          portfolio_item.update(quantity: (portfolio_item[:quantity].to_i - portfolio_item_params[:quantity].to_i))
+          render json: portfolio_item
+          return
+        else 
+          portfolio_item.destroy
+          render json: portfolio_item
+        end 
       end 
 
       private
